+++ src/app/journey/journey-result-row/result-trip-leg/result-trip-leg.component.ts
@@ -330,7 +330,7 @@ export class ResultTripLegComponent implements OnInit {
     let location = isFrom ? leg.fromLocation : leg.toLocation
 
     const locationData = <LegLocationData>{
-      locationText: location.locationName ?? '',
+      locationText: location.computeLocationName() ?? '',
       platformText: null,
       timeText: null,
       delayText: null,
@@ -357,8 +357,12 @@ export class ResultTripLegComponent implements OnInit {
         }
       }
 
-      locationData.platformText = stopPoint.plannedPlatform
-      locationData.actualPlatformText = stopPoint.actualPlatform
+      locationData.platformText = stopPoint.plannedPlatform;
+      if (stopPoint.actualPlatform === stopPoint.plannedPlatform) {
+        locationData.actualPlatformText = null;
+      } else {
+        locationData.actualPlatformText = stopPoint.actualPlatform;
+      }
     }
 
     return locationData